docs(gfql): keep the benchmark rules out of the docutils import
CI's minimal lane has no docutils, so putting the contract re-verification, the
staleness rule and the point-of-use decision in the Sphinx extension meant the
whole gate could only run in the docs build - one job out of forty, and the
pattern this repository has just finished digging out of.

`gfql_bench_data` is now stdlib-only and holds every rule; `gfql_bench` is the
docutils half that renders them and turns a recorded problem into a failed
build. A test asserts the rules module never grows a docutils or sphinx import
again.
